Revise the review comments and move the signal processor code location
